@@ -23,7 +23,7 @@
   },
   "homepage": "https://github.com/keokilee/react-electron-webview#readme",
   "dependencies": {
-    "camelcase": "^2.0.1"
+    "camelcase": "^4.0.0"
   },
   "devDependencies": {
     "babel": "^6.3.26",
@@ -32,19 +32,19 @@
     "babel-preset-es2015": "^6.3.13",
     "babel-preset-react": "^6.3.13",
     "babel-register": "^6.3.13",
-    "electron-prebuilt": "^0.36.2",
+    "electron-prebuilt": "^1.4.6",
     "expect": "^1.13.4",
-    "karma": "^0.13.16",
+    "karma": "^1.3.0",
     "karma-electron-launcher": "^0.1.0",
-    "karma-mocha": "^0.2.1",
+    "karma-mocha": "^1.3.0",
     "karma-sourcemap-loader": "^0.3.6",
-    "karma-spec-reporter": "0.0.23",
+    "karma-spec-reporter": "0.0.26",
     "karma-webpack": "^1.7.0",
-    "mocha": "^2.3.4",
-    "react": "^0.14.5",
-    "react-addons-test-utils": "^0.14.5",
-    "react-dom": "^0.14.5",
-    "webpack": "^1.12.9",
-    "semantic-release": "^4.3.5"
+    "mocha": "^3.1.2",
+    "react": "^15.3.2",
+    "react-addons-test-utils": "^15.3.2",
+    "react-dom": "^15.3.2",
+    "semantic-release": "^4.3.5",
+    "webpack": "^1.12.9"
   }
 }
